refactor(background): clarify persistence ownership
- move background investigation records and storage into a dedicated platform package
- shorten contract docstrings and document the project-wide guideline
- clean temporary files without catching BaseException
