+++ platform/background_investigations/__init__.py
@@ -1,0 +1,1 @@
+"""Durable background-investigation records shared across surfaces."""
